# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Movie Club API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me/avatar/upload-url:
    post:
      operationId: getAvatarUploadUrl
      tags: [auth]
      summary: Get a signed URL for uploading a profile picture
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetAvatarUploadUrlBody"
      responses:
        "200":
          description: Upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
  /me/avatar:
    patch:
      operationId: updateAvatar
      tags: [auth]
      summary: Update user's profile picture URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvatarBody"
      responses:
        "200":
          description: Avatar updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /movies/search:
    get:
      operationId: searchMovies
      tags: [movies]
      summary: Search movies via OMDB
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MovieSearchResult"
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List user's groups
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupSummary"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupBody"
      responses:
        "201":
          description: Created group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get group details
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: weekOf
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupId}/invite:
    get:
      operationId: getActiveInvite
      tags: [invites]
      summary: Get active invite code for group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Active invite or null if none exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActiveInvite"
    post:
      operationId: createInvite
      tags: [invites]
      summary: Create invite link/code
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteBody"
      responses:
        "201":
          description: Invite created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
  /groups/join:
    post:
      operationId: joinGroup
      tags: [invites]
      summary: Join group via invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinGroupBody"
      responses:
        "200":
          description: Joined group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          description: Invalid or expired invite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupId}/kick:
    post:
      operationId: kickMember
      tags: [groups]
      summary: Remove a member from the group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KickMemberBody"
      responses:
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /groups/{groupId}/role:
    post:
      operationId: updateMemberRole
      tags: [groups]
      summary: Update a member's role
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRoleBody"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /groups/{groupId}/picker:
    post:
      operationId: assignPicker
      tags: [groups]
      summary: Assign the weekly movie picker
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignPickerBody"
      responses:
        "200":
          description: Picker assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /groups/{groupId}/movie:
    post:
      operationId: setMovie
      tags: [movies]
      summary: Set the weekly movie
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetMovieBody"
      responses:
        "200":
          description: Movie set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieData"
  /groups/{groupId}/vote:
    post:
      operationId: submitVote
      tags: [votes]
      summary: Submit a rating for the current movie
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitVoteBody"
      responses:
        "200":
          description: Vote submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or already voted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupId}/watch-status:
    post:
      operationId: setWatchStatus
      tags: [groups]
      summary: Set watched/not-yet status for the current week
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetWatchStatusBody"
      responses:
        "200":
          description: Watch status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /groups/{groupId}/status:
    get:
      operationId: getGroupStatus
      tags: [groups]
      summary: Get voting status and current movie for a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: weekOf
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Group status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupStatus"
  /groups/{groupId}/results:
    get:
      operationId: getResults
      tags: [votes]
      summary: Get results for a movie week (only after deadline)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: weekOf
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Results"
        "403":
          description: Results not available yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /invites/{code}:
    get:
      operationId: getInvite
      tags: [invites]
      summary: Get invite details by code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invite details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupId}/turns/{turnIndex}/extension:
    post:
      operationId: setTurnExtension
      tags: [groups]
      summary: Set extension days for a specific turn (admin/owner only)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: turnIndex
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetTurnExtensionBody"
      responses:
        "200":
          description: Turn extension set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TurnExtensionResult"
  /groups/{groupId}/nominations:
    get:
      operationId: listNominations
      tags: [nominations]
      summary: List all nominations for a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Nominations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Nomination"
    post:
      operationId: submitNomination
      tags: [nominations]
      summary: Submit a movie nomination for a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitNominationBody"
      responses:
        "201":
          description: Nomination created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Nomination"
        "409":
          description: Already nominated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupId}/nominations/{nominationId}:
    delete:
      operationId: deleteNomination
      tags: [nominations]
      summary: Remove a nomination (admin/owner only)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: nominationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Nomination removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Sticker endpoints
  /stickers/upload-url:
    post:
      operationId: getStickerUploadUrl
      tags: [stickers]
      summary: Get a signed URL for uploading a sticker image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetUploadUrlBody"
      responses:
        "200":
          description: Upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
  /stickers:
    get:
      operationId: listGlobalStickers
      tags: [stickers]
      summary: List all global stickers
      responses:
        "200":
          description: Global stickers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlobalStickersResponse"
    post:
      operationId: createGlobalSticker
      tags: [stickers]
      summary: Create a global sticker (super admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStickerBody"
      responses:
        "201":
          description: Sticker created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sticker"
  /stickers/{stickerId}:
    delete:
      operationId: deleteGlobalSticker
      tags: [stickers]
      summary: Delete a global sticker (super admin only)
      parameters:
        - name: stickerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sticker deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteStickerResponse"
  /groups/{groupId}/stickers:
    get:
      operationId: listGroupStickers
      tags: [stickers]
      summary: List stickers available for a group (global + group-specific)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stickers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupStickersResponse"
    post:
      operationId: createGroupSticker
      tags: [stickers]
      summary: Create a group-specific sticker (admin only)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStickerBody"
      responses:
        "201":
          description: Sticker created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sticker"
  /groups/{groupId}/stickers/{stickerId}:
    delete:
      operationId: deleteGroupSticker
      tags: [stickers]
      summary: Delete a group sticker (admin only)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: stickerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sticker deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteStickerResponse"

  # Reaction endpoints
  /reactions:
    get:
      operationId: getReactions
      tags: [reactions]
      summary: Get reactions for an entity
      parameters:
        - name: entityType
          in: query
          required: true
          schema:
            type: string
        - name: entityId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reactions summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionsResponse"
    post:
      operationId: createReaction
      tags: [reactions]
      summary: Add a reaction to an entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReactionBody"
      responses:
        "201":
          description: Reaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reaction"
  /reactions/toggle:
    post:
      operationId: toggleReaction
      tags: [reactions]
      summary: Toggle a reaction (add if not exists, remove if exists)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleReactionBody"
      responses:
        "200":
          description: Reaction toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToggleReactionResponse"
  /reactions/details:
    get:
      operationId: getReactionDetails
      tags: [reactions]
      summary: Get detailed reactions (who reacted with what)
      parameters:
        - name: entityType
          in: query
          required: true
          schema:
            type: string
        - name: entityId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Detailed reactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionDetailsResponse"
  /reactions/{reactionId}:
    delete:
      operationId: deleteReaction
      tags: [reactions]
      summary: Remove a reaction
      parameters:
        - name: reactionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reaction removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    LoginBody:
      type: object
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 32
        password:
          type: string
          minLength: 8
          maxLength: 100
      required:
        - username
        - password
    RegisterBody:
      type: object
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 32
        password:
          type: string
          minLength: 8
          maxLength: 100
      required:
        - username
        - password
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        createdAt:
          type: string
        avatarUrl:
          type: ["string", "null"]
      required:
        - id
        - username
        - createdAt
    GetAvatarUploadUrlBody:
      type: object
      properties:
        filename:
          type: string
        contentType:
          type: string
      required:
        - filename
        - contentType
    UpdateAvatarBody:
      type: object
      properties:
        avatarUrl:
          type: string
      required:
        - avatarUrl
    MovieSearchResult:
      type: object
      properties:
        imdbId:
          type: string
        title:
          type: string
        year:
          type: string
        poster:
          type: ["string", "null"]
      required:
        - imdbId
        - title
        - year
        - poster
...
    CreateGroupBody:
      type: object
      properties:
        name:
          type: string
        startDate:
          type: string
          description: "ISO date (YYYY-MM-DD) when turns start. Defaults to today."
        turnLengthDays:
          type: integer
          minimum: 1
          maximum: 365
          description: "Length of each turn in days. Defaults to 7."
      required:
        - name
    Group:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
        ownerId:
          type: integer
        startDate:
          type: string
        turnLengthDays:
          type: integer
      required:
        - id
        - name
        - createdAt
        - ownerId
        - startDate
        - turnLengthDays
    GroupSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        memberCount:
          type: integer
        currentMovie:
          type: ["string", "null"]
        moviePoster:
          type: ["string", "null"]
        votingOpen:
          type: boolean
        hasVoted:
          type: boolean
        resultsAvailable:
          type: boolean
        turnLengthDays:
          type: integer
        startDate:
          type: string
      required:
        - id
        - name
        - role
        - memberCount
        - currentMovie
        - moviePoster
        - votingOpen
        - hasVoted
        - resultsAvailable
        - turnLengthDays
        - startDate
...
    CreateInviteBody:
      type: object
      properties:
        expiresInHours:
          type: integer
    Invite:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        groupId:
          type: integer
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - code
        - groupId
        - expiresAt
        - createdAt
    ActiveInvite:
      type: object
      properties:
        id:
          type: integer
        code:
          type: ["string", "null"]
        groupId:
          type: integer
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - code
...
    JoinGroupBody:
      type: object
      properties:
        code:
          type: string
      required:
        - code
    KickMemberBody:
      type: object
      properties:
        userId:
          type: integer
      required:
        - userId
...
    UploadUrlResponse:
      type: object
      properties:
        uploadUrl:
          type: string
        objectName:
          type: string
        publicUrl:
          type: string
      required:
        - uploadUrl
        - objectName
        - publicUrl
...